+++ iOS/IssueTracker/IssueTracker/IssueCreation/ViewModel/IssueCreationViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  IssueCreationViewModel.swift
+//  IssueTracker
+//
+//  Created by cho on 2020/11/10.
+//  Copyright © 2020 ralph. All rights reserved.
+//
+
+import Foundation
